DEVEL15-cmd-nname-20080113
LICENSE MIT

Nname() is used to concatenate two strings and is frequently used with
the first string being the name of the executable perhaps with a full
path.  The static buffer specified is too small for a full path and
there was no protection against writing beyond the end of it.


(cherry picked from commit d5811091995b78d65e891b134aa0ad6955bbc30c)