windows-freelance-improved-dfs-handling-20080127
LICENSE MIT

Cell names are published as share names.  As a result they are searched
for as part of the Dfs Referral evaluation.  If share "foo" can not be
reached, the CIFS client will fallback to searching for "fo" as well.

Since the freelance client automatically adds a symlink for prefixes
of the cell name, this results in both "foo" and "fo" being added to
the freelance root.afs volume.  This patch prevents that negative side
effect.