windows-lanahelper-20080128
LICENSE MIT

The algorithms that were used to produce a Netbios name were broken.
The name that was produced when the hostname was too long was in fact
longer than the maximum permitted netbios name.

Also, the "NetbiosName" value was not used as a suffix when the loopback
adapter was not installed.  The hardcoded string "AFS" was used instead.