DEVEL15-windows-lanahelper-20080128
LICENSE MIT

The algorithms that were used to produce a Netbios name were broken.
The name that was produced when the hostname was too long was in fact
longer than the maximum permitted netbios name.

Also, the "NetbiosName" value was not used as a suffix when the loopback
adapter was not installed.  The hardcoded string "AFS" was used instead.


(cherry picked from commit 2542c5e07f52c89f25f290153f0b4669076ae654)