windows-refactor-cm_config-20080201
LICENSE MIT

cm_config.c was a nightmare.  there were several functions that each
computed the location of the OpenAFS Client configuration directory
in a slightly different way.  Now there is a single exported function
that determines the directory, and another that determines the name
of the CellServDB file.  All of the other functons in cm_config.c and
throughout the src/WINNT/afsd directory now use those two functions.

The new rule for determining the location of the OpenAFS Client
configuration directory and hence the CellServDB file location is:

1. check the AFSCONF environment variable
2. check the HKLM\SOFTWARE\OpenAFS\Client   CellServDBDir registry value
3. use AFSDIR_CLIENT_ETC_DIRPATH which currently refers to
   \Program Files\OpenAFS\Client