windows-more-cellservdb-consolidation-20080201
LICENSE MIT

Merge cellconfig.c and the afsreg library.
Redefine the location of the client cellservdb directory in dirpath
in terms of the cellservdb direction and not the install directory.
(although they point at the same thing now by default.)

Future changes to the location of the cellservdb directory should
be made in afsreg.