windows-no-more-rpcns4-dll-20080212
LICENSE MIT

Remove all references to rpcns4.lib and its included functions RpcNsXXX.
These APIs are no longer supported on Vista.  Except for TaAfsAdmSvr
they had been commented out for a long time but we still continued to
link to the library.  Loading the library generates a no longer supported
error on Vista.  So lets get rid of it.