DEVEL15-windows-dfs-referrals-20080212
LICENSE MIT

Add "ipc$" to the list of names that must not be automatically added
as a share name.

If the sharename is a partial match and therefore will not be added
automatically to the share list, make sure that we return an error
to the requestor.


(cherry picked from commit 44c20384ed5c0fdb4bce737a2a3c2b9c0af4472b)