dafs-accurately-track-inuse-20080317
LICENSE IPL10

keep accurate tabs on whether a volume is "in use" so we don't end up with volume state that's nonsensical.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
LICENSE IPL10

keep accurate tabs on whether a volume is "in use" so we don't end up with volum
e state that's nonsensical.

====================
LICENSE IPL10

keep accurate tabs on whether a volume is "in use" so we don't end up with volum
e state that's nonsensical.