windows-aklog-robustness-20080330
LICENSE MIT

Fix the order of parameters to strncpy

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
LICENSE MIT

Fix the order of parameters to strncpy

Add error checking to all krb5 functions.

Initialize the global ccache variable to NULL.