DEVEL15-windows-freelance-20080411
LICENSE MIT

Do not set the DV on the rootSCachep because doing so disables
the use of MergeStatus().

Do not remove the rootSCachep from the hash table because doing
so makes causes a second cm_scache_t object for that fid to be created
that is not the root.


(cherry picked from commit 7bfe889eb2e900719c70cd9fd29f82569ef4c04c)