KDFS-windows-volume-status-20080418
LICENSE MIT

A problem was discovered with cm_Analyze when serverRef instances are
marked offline.  cm_CheckOfflineVolume() was not resetting the serverRef
state.  cm_Analyze was also waiting in sleep calls when it was it is
clear that waiting would not result in a change of state.

cm_CheckOfflineVolume() was updated to always reset the serverRef state
and indicate that the volume should be considered online for the purpose
of retrying when the serverRef state is reset to srv_not_busy.

The problems identified in the MIT dev.mit.edu stress test environment
are resolved by these changes.


(cherry picked from commit c3eb805fc0309efca059a26b946eb4350417867e)