macos-afscell-configurator-20080423
LICENSE IPL10

not ready for release, but, add a tool which installer can run to ask for a cell
name
AUTHOR
LICENSE IPL10

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
LICENSE IPL10

not ready for release, but, add a tool which installer can run to ask for a cell
name
(cherry picked from commit b1bf96f7581e58137f49e2b685cf6e91ff4f7dda)

Change-Id: I13ca9268e679e029f97260712f1ac7331fd90b2d
Reviewed-on: http://gerrit.openafs.org/916
Tested-by: Derrick Brashear <shadow@dementia.org>
Reviewed-by: Derrick Brashear <shadow@dementia.org>