macos-afscell-configurator-20080423
LICENSE IPL10

not ready for release, but, add a tool which installer can run to ask for a cell
name
AUTHOR
LICENSE IPL10

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
LICENSE IPL10

not ready for release, but, add a tool which installer can run to ask for a cell
name