linux-hlist-unhashed-opencoding-20080520
LICENSE IPL10
FIXES 93871

if we can't use hlist_unhashed (which is dumb) or we don't have it (less dumb) work around it.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
LICENSE IPL10
FIXES 93871

if we can't use hlist_unhashed (which is dumb) or we don't have it (less dumb) w
ork around it.