Name the kernel module openafs instead of libafs
Apply quilt patch module-name: Build the kernel module as openafs.ko
instead of libafs.ko.  This should really be done upstream as well, but it
poses backwards compatibility problems and it's always been called libafs
on other platforms, so it's controversial.

This implementation of the idea produces warnings and needs some
improvement.