Ugly hack to build the PAM modules PIC
Apply quilt patch pam-build: Apply an incredibly ugly hack to build the
PAM modules against the shared objects generated for the shared libraries.
Not suitable upstream in its current form.  The right upstream solution is
to standardize the shared library ABI and install it with a real SONAME,
and then link the PAM modules against it.  It may be that we'll drop
kaserver support and stop building the PAM modules before we get to that,
though.