Tweak the expected configuration directory permissions
Apply quilt patch bos-permissions: Expect the server directory to be
mode 0700 since that's where the AFS KeyFile is stored.  Allow the
local directory to be mode 0755, since there's no reason for it to be
more restrictive.

Allowing the server directory to be mode 0700 causes problems upstream if
people symlink the client configuration to the server directory, which is
recommended by some documentation, so this is not applied upstream.