Allow overriding the default compiler and merge SPARC
Apply quilt patch compiler-flags: Use $CC for the compiler instead of
forcing cc or gcc.  The corresponding upstream fix is waiting for work
to separate the kernel compiler from the userspace compilter, since $CC
should only be used for the latter on other platforms.

Also combines the sparc sysnames together, which is suitable upstream.