Allow multiple local realms in krb.conf
Apply quilt patch multiple-local-realms: Enable listing multiple realms
in the server krb.conf file.  Up to four realms may be treated as local
in that fashion for authorization purposes (instead of the single realm
supported without this patch).

This upstream patch is in the 1.5 series but not in 1.4.  However, it
has been working in production at Stanford University for some time.