KDFS-windows-afsrdr-build-20080702
LICENSE MIT

Add src/WINNT/afsrdr/ to the build system.  afsrdr/NTMakefile constructs
two command files, ddkbuild.cmd and ddkclean.cmd, that instantiate a DDK
build environment and within that environment perform the appropriate
build and clean operations.

The output of the DDK build environment is afsrdr/build/...

Code signing requires the DDK version of signtool and the addition of
a cross-signed certificate that is embedded into the driver.  A new
environment variable CODESIGN_CROSS_CERT provides the path to the
certificate to be embedded.  See for details:

http://www.microsoft.com/whdc/winlogo/drvsign/kmsigning.mspx
http://www.microsoft.com/whdc/winlogo/drvsign/crosscert.mspx