DEVEL15-windows-install-loopback-ddk-20080710
LICENSE MIT

Add support for dynamically determining which ddk include
path should be used.  The NTDDKDIR variable used to be set
to the specific directory in which netcfgx.h was located.
Now it is set to the top-level of the DDK install tree.
The specific directory that contains netcfgx.h varies based
upon whether the DDK is 3790.1830 or 6000.  Test for file
existence and specify the one that works.


(cherry picked from commit 4f5b808414c958e7629c83f93d610b57c54274ac)