windows-smb-misc-20080714
LICENSE MIT

Fix declaration of smb_ParseStringBuf to be consistent with
the prototype.

If the fileType of a cm_scache_t is unknown and if the vnode is odd,
then set the SMB_ATTR_DIRECTORY flag when reporting the DOS Extended
Attributes.

Fix the log message in smb_FindCMUserByName() to indicate "found"
instead of "not found".