Drop support for hppa
We only have a param file for a 2.4 kernel, we were building with UCONTEXT
which is stubbed out in glibc for hppa, and builds without UCONTEXT fail.
The binaries we were building before didn't work and no one reported a
bug, indicating no one was attempting to use OpenAFS on hppa.