DEVEL15-windows-ioctl-prefix-data-20080726
LICENSE MIT

The ioctl-prefix-data can be accessed either via a char_t or a wchar_t.
Copy the data buffer with memcpy() and not strcpy() because we need to
copy the entire string if it happens to be Unicode.


(cherry picked from commit eb1ff11bf14cfc64b4e224ef574d62e6e3757bc1)