KDFS-windows-large-files-20080728
LICENSE MIT

as part of the restructuring of code to separate the smb layer
from the cm layer, large file support was broken.   Define AFS_LARGEFILES
in the right place so that it will be used in the cm.


(cherry picked from commit b45b059ed9772642489d56372cbd74db4f0cfc02)