KDFS-windows-callback-shutdown-head-20080801
LICENSE MIT

When the cache manager is shutting down, return an error to all
callback requests.  Otherwise, it is possible for a callback to get
stuck waiting for a lock and when the lock clears to be left with
nothing but invalid memory addresses to crash on.

In the meantime, the file server is waiting for the response.


(cherry picked from commit 25b819543ea4343216b6a2955396719a49ea042f)