rename-rx-stats-20080925
LICENSE IPL10

avoid using the same name for the struct and the global so the windows debugger works usefully

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
LICENSE IPL10

avoid using the same name for the struct and the global so the windows debugger
works usefully

====================
LICENSE IPL10

avoid using the same name for the struct and the global so the windows debugger
works usefully