smb-symlink-to-vnovnode-attribute-20081017
LICENSE MIT

When reporting the attribute of a directory name which happens to be
a symlink object to a file that cannot be accessed or does not exist,
use the target name to guess the type of the object.  If it has an
extension, consider it a file, otherwise report it as a directory.