windows-afsd-hook-20081106
LICENSE MIT

When loading the afsdhook.dll require that it be in the same directory
as afsd_service.exe.  This prevents LoadLibrary() from searching the PATH
environment variable if it cannot be found.   If the PATH environment
variable contains a path in AFS, the library will fail to load and take
a very long time doing so.