KDFS-rdr-tools-20081107
LICENSE BSD

SetTrace:

This application sets the parameters for the tracing. They can be:

    /l - Logging level (\Parameters\DebugLevel)
        1: Error
        2: Warning
        3: Verbose
        4: Verbose 2

    /s - Subsystem  These values can be OR'd together
         (\Parameters\TraceSubsystem)
        1: IO Subsystem
        2: File subsystem
        4: Lock processing, level must be set to Verbose
        8: Extent processing

    /b - Trace buffer size in KB (Set this to zero to turn off tracing)
         (\Parameters\TraceBufferSize)

When you change settings there is no requirement to reboot then system.
The corresponding registry values are updated automatically. Do not set
a value if you don't want to change it's current value.

If you want the trace statements to be printed to the debugger set the
DebugFlags value in the file system Parameters key to 2.


GetTrace:

This application retrieves the entire trace buffer from the file system

Crash:

This will crash the system.