KDFS-windows-rdr-user-symlink-parents-20081108
LICENSE MIT

When evaluating the target of a symlink the file system may not know
who the symlink's parent is.  Especially if it is evaluating the target
by the FID.  The symlink's parent can be constructed from the symlink's
cell/volume combined with the ParentVnode/ParentUnique fields from
the cm_scache_t.  This will permit evaluation under all circumstances.