KDFS-windows-afsrdr-common-20081119
LICENSE BSD

define service to fs request flags.  AFS_REQUEST_RELEASE_THREAD
adds the specified thread to a pool of workers that can only be
used for operations that release extents to the service.

add Maximum Chunk Length and a Hide Dot Files flag to
AFSCacheFileInfo; renamed to AFSRedirectorInitInfo

Add Filesystem attribtes to AFSVolumeInfoCB

Add a new subsystem to the trace logic: worker processing

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
LICENSE BSD

define service to fs request flags.  AFS_REQUEST_RELEASE_THREAD
adds the specified thread to a pool of workers that can only be
used for operations that release extents to the service.

add Maximum Chunk Length and a Hide Dot Files flag to
AFSCacheFileInfo; renamed to AFSRedirectorInitInfo

Add Filesystem attribtes to AFSVolumeInfoCB

Add a new subsystem to the trace logic: worker processing

Remove many deadlocks and race conditions

Add lots of trace logging

Reimplement directory invalidation logic.  The file system does not
tear down the directory or the child FCBs when the directory is marked
invalid.  Instead a new directory enumeration is performed and the
difference of the new and old directory is used to determine what should
be torn down or created.

Obtain Volume Info from the service.  Volume name is reported as
the first 20 characters of the AFS volume name and the serial number
is set to the AFS volume ID.  Partition size and free space are used.

Fix the shutdown logic to block all incoming requests while the shutdown
is in process, complete outstanding requests, return all extents to
service, and after service disconnect clear all flags.

Implement reparse logic for DFS links.

Trim down extents prior to reporting the truncation of a file to the
service to avoid a race.

Permit special share names \\afs\cell{#,%}volume\ to be processed.

Implement logic to support "fs flushxxx" operations.

Clear the Windows system cache when objects are invalidated.

Block on the last handle close for each process and flush all dirty
extents to the service.

Issue a Set File Info (length) when a file is being truncated.