KDFS-windows-afsrdr-service-20081119
LICENSE MIT

implement two phase shutdown for the file system

conditionalize all file system operations with RDR_Initialized

avoid unnecessary thrashing of the Btree construction if the dir lookup
operation began with an older version of the directory.

add buf_rdrReleaseExtentsLock to prevent multiple simultaneous
attempts to request extents from the file system.

add buf_FindAll() and buf_FindAllLocked() to permit buffers to be
searched for on the all buffer list

add cm_req_t objects to buf_Getxxxx() functions to permit file system
operations to be non-blocking.  If a buffer cannot be obtained return
CM_ERROR_WOULDBLOCK when CM_REQ_NORETRY is also specified.

Wherever a buffer cannot be touched because it is dirty, it also cannot
be touched when the file system owns it.

Add buffer ref counting for the file system ownership

Add cm_req_t to cm_MergeStatus() to permit file system specific
processing.  Do not send a data invalidation if the cause of the
invalidation was a file system request.

Raise the default number of daemons to 8

Raise the max number of daemons to 64

Add a 5 second sleep after a VBUSY prior to calling cm_FindVolumeByID
to regenerate the service list.  This is to avoid problems caused by
volumes being moved on buggy servers that take the volume offline
during the operation instead of holding the client in an idle data
sequence.

Add cm_REQ_SOURCE_REDIR

Handle error responses when attempting to invalidate objects held
by the file system.

Do not implicitly release extents if the associated operation fails.

Deliver volume status messages in a thread independent of the
operation that resulted in the volume state change being detected.
This is to avoid deadlocks where the file system holds a lock,
calls the service and the volume state change processing requires
the lock that was already being held.  (this implementation needs
to be revisited.)