KDFS-windows-afsd-redir-20081208
LICENSE MIT

Conditionalize all calls to RDR_InvalidateObject with
an RDR_Initialized test.  This is done to reduce the
performance hit when running with the SMB interface.

Add BUF_ERROR_TRUNC (-2) to cm_buf which is used for
debugging purposes to indicate that a cm_buf_t's
dataVersion field was set to CM_BUF_VERSION_BAD due
to a truncation as opposed to an actual error code.

in cm_dcache do not attempt to release a cm_buf_t
when buf_Get() fails.

Do not issued RDR_InvalidateObject(DELETED) if the
source of the deletion was the file system.