KDFS-windows-rdr-user-request-file-extents-20081208
LICENSE MIT

Change RDR_RequestFileExtentsAsync to return a BOOL
which indicates whether or not the request should
be retried.  When a wouldblock condition is reached
instead of simply returning the extents that have
been obtained so far and ignoring the rest, return
what we have obtained and retry the remaining extents.
This will avoid the file system from being forced to
reissue the request.

If we subsequently receive a duplicate request for
an extent that has already been given to the file
system that is ok.  It will be ignored.