KDFS-windows-afs-rdr-common-20081208
LICENSE BSD

Add QueueCount to the AFSCommRequestBlock.
This number is an estimate of the number of
outstanding requests the file system has
to send to the afs cache manager.  This
value can be used in the future to dynamically
adjust the worker thread pool size based upon
the observed need.