KDFS-windows-afs-rdr-kernel-20081208
LICENSE BSD

(jaltman@secure-endpoints.com contributed
to these changes.)

Add QueueCount to the AFSCommRequestBlock.
This number is an estimate of the number of
outstanding requests the file system has
to send to the afs cache manager.  This
value can be used in the future to dynamically
adjust the worker thread pool size based upon
the observed need.

Add debugging to AFSEnumerateDirectory.

Add assertions to ensure that VolumeInformation
is only accessed for Fcbs that are root volume
objects (vnode = 1, unique = 1).

The AFS Global Root Fcb is a root volume.  Set
the vnode,unique pair to indicate that.

If an AFSEnumerateDirectory call is interrupted,
reset the directory contents.

Do not initialize the DirtyEntentsList without
holding the lock resource.

When waiting for an extent, periodically timeout
the wait and re-issue the request to the cache
manager.  There is a potential race in low extent
count scenarios where an extent can be assigned
to the file system and released back to the AFS
cache manager before the requesting thread has
a chance to use the required extent.

Use IsListEmpty() instead of relying on the
return value to RemoveHeadList() to indicate
that the list is empty.

Reorder the acquisition of the NPDirNode and Fcb
locks in AFSValidateEntry to avoid a deadlock.