KDFS-windows-afsrdr-kernel-20081210
LICENSE BSD

During a rename operation if the target name is in use
and the replace on rename flag is set, replace the previous
target instead of failing the operation.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
LICENSE BSD

(assisted by pscott@kerneldrivers.com)

Release extents when they have been idle for 6 minutes.

Do not permit AFSLocateNameEntry to return
STATUS_OBJECT_NOT_FOUND from called function.  This
error has special meaning when returned from this
function.

In AFSBuildTargetDirectory, do not treat non-MountPoints
as if they are MountPoints

Cleanup and addition of comments.

====================
LICENSE BSD

AFSInitRootForMountPoint was originally intended to be called
only to evaluate mount point targets.  However, it is now being
called to evaluate SymbolicLink targets as well.  As a result
we cannot count on the vnode being equal to 1 (indicating a
volume root) instead we must set the vode to be 1.  This permits
us to construct an Fcb for the target of a SymbolicLink that
might be in a volume the file system has never encountered before.