KDFS-windows-afsrdr-user-20081210
LICENSE MIT

do not walk the entire list of extents during a release call
trying to find dirty ones to flush if we have already flushed
them all.

Always call cm_CheckNTOpenDone on success or failure of
cm_CheckNTOpen.

Cleanup logging statements.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
LICENSE MIT

Permit _WIN32_WINNT to be defined by the build system.  Default to 0x0500
if it is unspecified.

In RDR_EnumerateDirectory(), if the ResultBufferLength is 0 there is
no room to respond with a AFSDirEnumResp structure.   Therefore,
conditionalize all assignments to that structure based upon the
ResultBufferLength.

Fix the call to cm_Gen8Dot3VolNameW which takes four arguments
not three.