KDFS-windows-afsrdr-kernel-20081212
LICENSE BSD

Enumerate the global root immediately after afsd_service initializes
the file system.  Do not wait for the first request to be received.

The pioctl interface queries the ready state of the file system before
issuing a request.  If the global root has not been enumerated, the
file system will report a not ready status and the pioctl will fallback
to the smb interface.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
LICENSE BSD

change the network provider comment for the "AFS" server
entry to "AFS Root".

return success if the invalidation cannot find the object