KDFS-windows-afsrdr-npdll-20081212
LICENSE BSD

Obtain network provider name from the registry.  This ensures
consistency and permits site specific configuration.

Upper case local names for consistency.

When enumerating only return the number of entries requested
by the caller.

In NPGetResourceInformation make sure the lpComment and lpProvider
are initialized.  Do not return the comment but be sure to return
the provider.