KDFS-windows-afsrdr-kernel-20081215
LICENSE BSD

Add FID to all trace messages that include a DirEntry

Fix spelling errors in comments

Make sure that the PrivateCacheMap, created with
CcInitializeCacheMap(), is still in place prior to issuing
calls to Cc API functions.  If the PrivateCacheMap was
released, reconstruct it.

In AFSVolumeWorkerThread, do not block when attempting to
voluntarily release extents.  Doing so can result in a
multiple resource chain deadlock and severely hampers
performance.