KDFS-windows-afsrdr-npdll-20081215
LICENSE BSD

Add missing export for GetConnection3

Replace the dynamically allocated with szProviderName with a static
buffer in order to avoid memory leaks. The DllMain() is not being
called by the C RTL.

In GetResourceInformation, the output strings must be NUL terminated
and the input buffer size must be preserved so it can be used to
determine if there is sufficient room or not.  Bad things happen
if we always return a "more data" error.