KDFS-windows-afsrdr-kernel-20081216
LICENSE BSD

when evaluating a relative path where the last component does not
exist, do not release the parent resource twice.

when enumerating directories, evaluate the symlink to determine the
type of the final target in the chain.  however, if the next to final
target is a mount point, do not evaluate the mount point as that is
an unnecessary expense.