KDFS-windows-afsrdr-kernel-np-support-20081217
LICENSE BSD

Modify the reported connection scope, type, usage values to make
them consistent with what is actually reported for dfs shares.

Validate requests for connection info for \\afs\cell#volume\
share names and create connection entries for those that exist.

With \\afs\cell#volume\ entries it is no longer a valid assumption
that only symlinks and mount points will be found as children of
the root.  (Actually this was never true.)

Do not attempt to remove an entry from the short name tree if
the tree is NULL.