KDFS-windows-afsrdr-kernel-20081219
LICENSE BSD

in collaboration with pscott@kerneldrivers.com

rename and repurpose AFSBuildTargetDirectory() to
AFSBuildMountPointTarget().  The function is now only
called to build the the target Fcb of a mount point.

AFSBuildSymLinkTarget() is called to evaluate the target
Fcb of a SymLink.  The target Fcb can be any type of
object.  Mount Point Fcbs are not automatically evaluated.

The SymbolicLink.TargetFcb field has been removed.
SymLink targets must be evaluated on each access in order
to ensure proper target resolution as intermediary path
objects can be altered at any time.

Restructure AFSParseName and AFSLocateName to improve
readability.