KDFS-windows-afsrdr-kernel-file-attributes-20081220
LICENSE BSD

New function AFSGetFileAttributes() is now used to obtain or
compute the FileAttributes that should be used in responses
to directory enumerations and information queries.

To improve readability replace FileType 0 with the descriptive
AFS_FILE_TYPE_UNKNOWN preprocessor symbol.

The DirEntry.DirectoryEntry.FileAttributes field is now
strictly the FileAttributes value reported by the service.
The file system should not modify this value.