KDFS-windows-afsrdr-kernel-extents-20081221
LICENSE BSD

If a read request is received that is larger than the afs cache
manager's maximum rpc length, break the request up into chunks
of maximimum rpc length before contacting the service.  This
permits the file system to begin using the received chunks while
the rest are being fetched.

Add logic to AFSMarkDirty to validate that all of the extents that
we are trying to mark dirty have actually been found.