windows-afsd-rename-collision-20081226
LICENSE MIT

Do not permit cm_Rename() to rename an existing object such that it
creates a case-insensitive name collision with another object.
This change in behavior is consistent with the restrictions when
creating new objects.  This change prevents the explorer shell from
getting confused when it creates a new folder.