KDFS-windows-install-wix-20081229
LICENSE MIT

Replace the default install with an IFS install and a SMB install.
The IFS install does not include the loopback adapter, afscreds, or
afs_config.

The SMB install does not include the file system driver or the
matching network provider.

Create new properties to permit a transform to easily select which
should be the default.  IFS is the default.

When generating the 32-bit tools installer, do not register the
network provider.   The configuration from the 64-bit install will
suffice for both versions.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
LICENSE MIT

remove references to IBM AFS release notes.