KDFS-windows-afsrdr-npdll-net-error-20081230
LICENSE BSD

when the redirector driver is not available or not ready, do not
return WN_NO_NETWORK as an error from the network provider.  doing
so results in the enumeration of all network providers to be halted
since what is the point if there is no network.

Instead return WN_NET_ERROR which indicates that something went
wrong for this provider.