KDFS-windows-afsrdr-kernel-build-symlink-target-20090109
LICENSE BSD

When the target of a symlink is a volume root, we must return
the control block of the volume root and not the current fcb
which is still referring to the last symlink in the chain.

Be sure to hold the TreeLock while we perform the is volume
test in order to prevent the volume control block from being
torn down.