rx-windows-xmit-20090126
LICENSE MIT

The WinSock API does not include the Posix recvmsg/sendmsg interfaces.
Beginning with XP/2003 Microsoft began to support WSARecvMsg which is
a mostly compatible implementation of recvmsg.  In Vista/2008 Microsoft
began to support WSASendMsg a mostly compatible implementation of sendmsg.

Neither are part of the WinSock API and therefore they must be loaded
at runtime via a WSAIoctl() call to obtain the function pointers.
When the functions are available it is now possible to avoid a large
number of memcpy() calls.

This patch also enables UDP port unreachable messages on XP and above.