windows-tests-nm-20090130
LICENSE BSD

A test application that produces verifiable random test files.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
LICENSE BSD

make it build in the openafs build tree