windows-tests-torture-enforce-querypath-info-errors-20090201
LICENSE BSD

change nb_qpathinfo() 'Flag' to a thread local storage variable
'EnforcePathInfoErrors'.

Add new 'Type' values that permit the enforcement to be turned on and off
as needed by the test script.

Populate the IoBuffer with the low byte of the pid value.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
LICENSE BSD

Verified the state of every file's existence at every Query Path Info
call in the script.  Added 'Type' == 1 (Does not Exist) as needed.
Removed the use of 'Type' == 1111 to disable enforcement part way through.