man-page-salvager-fixes-20090201
LICENSE BSD

Fix the -parallel example in the main description text: "5all" should be
"all5".  Reword the description a bit to hopefully make it clearer that
there are two separate values set here.  Note under the option description
of -parallel that multiple partitions on the same device are normally
processed serially.