KDFS-windows-afsrdr-common-20090313
LICENSE BSD

add HOLD_FID / RELEASE_FID interfaces.  Used by the file system to obtain
a reference count on a cm_scache_t object to ensure that it is not recycled.

Add GlobalFileId to the file system initialization arguments.  The file
system needs to be able to identify which volume is the global root
by fid in case the volume is mounted in multiple locations.