KDFS-windows-afsrdr-kernel-20090313
LICENSE BSD

with assistance from jaltman@secure-endpoints.com

Add GlobalFileId to the file system initialization arguments.  The file
system needs to be able to identify which volume is the global root
by fid in case the volume is mounted in multiple locations.

Reimplement all of the path parsing logic in order to be able to handle
symlinks that contain a ".." entry that crosses above a volume root.
The full starting path is now recorded in the Ccb as an array of component
names.  This is used as a stack pushing and popping components while
evaluating the contents of a symlink.

Permit directory enumeration to be performed asynchronously.

Correct spelling errors in comments.

Correct @sys processing.

When evaluating the target type for a symlink, do not evaluate the target
of the final component if that component is a mount point.

Initialize sysname resources.

Ensure that buffers are nul terminated before passing them to RtlRegQuery.

When processing the invalidation of a file deletion, delete the direntry
but not the fcb.  No longer assume that Fcb->DirEntry is always non-NULL
except during cleanup.