vol-fssync-server-socket-20090323
LICENSE IPL10

GetHandler() sets up a fdset and returns the max socket fd
which is used with select().  Windows ignores the numfd param
(first param) of select() and therefore computing the max fd
value is meaningless.  We can avoid a number of casts by
not attempting to treat this value as osi_socket.