vol-lock-fd-type-20090323
LICENSE IPL10

the lock_fd field of DiskPartition[64] is a file descriptor.
On Windows this is a HANDLE and on *nix platforms an int.
OpenAFS uses the FD_t type to provide platform specific type
info for file descriptors.  Use it for the lock_fd field and
the salvageLock in ObtainsSalvageLock().

Finally, the on the wire diskPartition[64] struct in volser/volint.xg
also contains a lock_fd field.  This is an on the wire field and
must be left at a fixed width of 32-bits.  Since a file descriptor
is not portable across machines we truncate the 64-bit HANDLE value
to fit in the 32-bit lock_fd field when necessary.